Delegate soltion to different files
